wip: draft nats client changes

+++ conftest.py
@@ -8,27 +8,57 @@ from test.mock_cases import cases as mock_cases
 from test.mock_cases import invalid_cases as mock_invalid_cases
 from test.parse_cases import cases as parse_cases
 from copy import deepcopy as dc
-from test.util import increment, get_lib_modules, get_nats_client, ping, startup_check
+from test.util import increment, get_lib_modules, ping, startup_check
 import functools
 import operator
 import json
 import importlib
+import nats
+
+def pytest_addoption(parser):
+    parser.addoption('--natscredfile', action='store', default=None)
 
 @pytest.fixture(scope='session')
-def nats_server():
+def nats_server(loop, request):
+    nats_cred_file = request.config.getoption('--natscredfile')
+
     print('Creating NATS fixture')
-    p = subprocess.Popen(['nats-server', '-D'])
-    yield p
-    p.terminate()
-    p.wait()
+
+    if nats_cred_file:
+        nats_url = 'tls://connect.ngs.global'
+    else:
+        nats_url = 'nats://127.0.0.1:4222'
+        p = subprocess.Popen(['nats-server', '-D'])
+    
+    yield nats_url, nats_cred_file
+    
+    if p:
+        p.terminate()
+        p.wait()
+
+
+@pytest.fixture(scope='session')
+def nats_client(loop, nats_server):
+    (url, cred_file) = nats_server
+
+    client = None
+
+    async def f():
+        nonlocal client
+        client = await nats.connect(url,  user_credentials=cred_file)
+
+    loop.run_until_complete(f())
+
+    yield client
 
 @pytest.fixture(scope='session', params=get_lib_modules())
-def dispatcher_server(loop, nats_server, request):
+def dispatcher_server(loop, nats_server, request, nats_client):
+    (nats_url, nats_cred_file) = nats_server
     lib_name = request.param
     test_module_name = 'lib.{}.dispatch'.format(lib_name)
     l = importlib.import_module(test_module_name)
 
-    s: subprocess.Popen = l.start()
+    s: subprocess.Popen = l.start(nats_url, nats_cred_file)
 
     try:                                
         startup_check(loop, lambda: ping(lib_name), times=20)
@@ -38,7 +68,6 @@ def dispatcher_server(loop, nats_server, request):
     yield s
 
     async def t2():
-        nats_client = await get_nats_client()
         req = json.dumps([{}, {'End': {}}])
         await nats_client.request(lib_name, req.encode(), timeout=1)
 
@@ -49,6 +78,8 @@ def dispatcher_server(loop, nats_server, request):
     except subprocess.TimeoutExpired:
         s.terminate()
         s.wait()
+
+
 
 # @pytest.fixture(scope='session', autouse=True)
 # def event_loop():
